list comprehension
This repo prefers comprehensions over maps as discussed in "What's new in Python 3" and CONTRIBUTING.md.
